Move example/test implementations to hotshot_testing
Closes #2059
