+++ crates/hotshot/Cargo.toml
@@ -148,3 +148,4 @@ blake3 = { workspace = true }
 clap = { version = "4.4", features = ["derive", "env"] }
 serde_json = "1.0.108"
 toml = { workspace = true }
+hotshot-testing = { path = "../testing" }
