+++ crates/hotshot/examples/combined/orchestrator.rs
@@ -2,7 +2,7 @@ pub mod types;
 
 use async_compatibility_layer::logging::{setup_backtrace, setup_logging};
 use clap::Parser;
-use hotshot::demo::DemoTypes;
+use hotshot_testing::demo::DemoTypes;
 use tracing::instrument;
 use types::VIDNetwork;
