+++ crates/hotshot/examples/libp2p/all.rs
@@ -5,9 +5,9 @@ use crate::types::ThisRun;
 use async_compatibility_layer::art::async_spawn;
 use async_compatibility_layer::logging::{setup_backtrace, setup_logging};
 use clap::Parser;
-use hotshot::demo::DemoTypes;
 use hotshot_orchestrator::client::ValidatorArgs;
 use hotshot_orchestrator::config::NetworkConfig;
+use hotshot_testing::demo::DemoTypes;
 use hotshot_types::traits::node_implementation::NodeType;
 use std::net::{IpAddr, Ipv4Addr};
 use tracing::instrument;
