+++ crates/hotshot/examples/libp2p/multi-validator.rs
@@ -3,8 +3,8 @@ use async_compatibility_layer::{
     logging::{setup_backtrace, setup_logging},
 };
 use clap::Parser;
-use hotshot::demo::DemoTypes;
 use hotshot_orchestrator::client::ValidatorArgs;
+use hotshot_testing::demo::DemoTypes;
 use std::net::IpAddr;
 use tracing::instrument;
 use types::VIDNetwork;
