+++ crates/hotshot/examples/libp2p/validator.rs
@@ -1,6 +1,6 @@
 use async_compatibility_layer::logging::{setup_backtrace, setup_logging};
 use clap::Parser;
-use hotshot::demo::DemoTypes;
+use hotshot_testing::demo::DemoTypes;
 use tracing::{info, instrument};
 use types::VIDNetwork;
