+++ crates/hotshot/examples/webserver/multi-webserver.rs
@@ -6,7 +6,7 @@ use async_compatibility_layer::{
     logging::{setup_backtrace, setup_logging},
 };
 use clap::Parser;
-use hotshot::demo::DemoTypes;
+use hotshot_testing::demo::DemoTypes;
 use tracing::error;
 
 #[derive(Parser, Debug)]
