+++ crates/testing/Cargo.toml
@@ -14,12 +14,18 @@ slow-tests = []
 
 [dependencies]
 ark-bls12-381 = { workspace = true }
+ark-serialize = { version = "0.3", features = [
+    "derive",
+] } # TODO upgrade to 0.4 and inherit this dep from workspace https://github.com/EspressoSystems/HotShot/issues/1700
+espresso-systems-common = { workspace = true }
 async-compatibility-layer = { workspace = true }
 async-trait = { workspace = true }
 # needed for vrf demo
 # so non-optional for now
 blake3 = { workspace = true, features = ["traits-preview"] }
+sha3 = "^0.10"
 commit = { workspace = true }
+derivative = "2.2.0"
 either = { workspace = true }
 futures = { workspace = true }
 hotshot = { path = "../hotshot", features = [
@@ -29,6 +35,7 @@ hotshot-types = { path = "../types", default-features = false }
 hotshot-utils = { path = "../utils" }
 hotshot-constants = { path = "../constants" }
 hotshot-orchestrator = { version = "0.1.1", path = "../orchestrator", default-features = false }
+hotshot-signature-key = { path = "../hotshot-signature-key" }
 hotshot-task = { path = "../task", version = "0.1.0", default-features = false }
 hotshot-task-impls = { path = "../task-impls", version = "0.1.0", default-features = false }
 jf-primitives = { workspace = true }
