+++ crates/types/src/lib.rs
@@ -12,7 +12,6 @@
 use displaydoc::Display;
 use std::{num::NonZeroUsize, time::Duration};
 use traits::{election::ElectionConfig, signature_key::SignatureKey};
-pub mod block_impl;
 pub mod consensus;
 pub mod data;
 pub mod error;
@@ -24,6 +23,7 @@ pub mod simple_vote;
 pub mod traits;
 pub mod utils;
 pub mod vote;
+
 /// the type of consensus to run. Either:
 /// wait for a signal to start a view,
 /// or constantly run
